requirements.txt
pytest
Appium-Python-Client
selenium==4.9.0